+++ docs/CHANGELOG.md
@@ -6,7 +6,7 @@ The format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and 
 
 ## [Unreleased]
 
-## [2.2.0] - 2026-07-28
+## [2.2.0] - 2026-07-29
 
 ### Added